+++ src/app/views/PersonDetail.svelte
@@ -62,7 +62,7 @@
   <div class="flex gap-4 text-gray-1">
     <PersonCircle {pubkey} class="mt-1 h-12 w-12 sm:h-32 sm:w-32" />
     <div class="flex min-w-0 flex-grow flex-col gap-4">
-      <Anchor class="flex flex-col" href={`/${npub}`}>
+      <div class="flex flex-col">
         <div class="flex items-center justify-between gap-4">
           <PersonName class="text-2xl" {pubkey} />
           <div class="hidden xs:block">
@@ -70,7 +70,7 @@
           </div>
         </div>
         <PersonHandle {pubkey} />
-      </Anchor>
+      </div>
       {#if $person.profile?.website}
         <Anchor
           external